Move component_wrappers.dart to utils folder
