+++ lib/src/context/message_context.dart
@@ -18,7 +18,7 @@ import 'package:nyxx_interactions/src/events/interaction_event.dart';
 
 import '../commands.dart';
 import '../commands/message_command.dart';
-import 'component_wrappers.dart';
+import '../util/component_wrappers.dart';
 import 'interaction_context.dart';
 
 /// Representsa  context in which a [MessageCommand] was executed.
